sql, roachpb: save representative logical plans in statement stats
Every `n` times a query with a particular fingerprint is run (currently
n is hardcoded to 1000), convert its logical plan to a simplified form
suitable for serialization to JSON or protobuf, and save that plan in
memory as part of the StatementStatistics struct.

Since the admin server's `Queries` method returns the full, unscrubbed
StatementStats proto for each fingerprint, this change exposes these
plans to the admin UI.

A future commit could scrub the plans so they can be sent to the reg
cluster.

Release note: None